Extract sample test classes from StackTracesTest
Follow Kevin's idea from ead2fe7ca97e4b0bc013efe0e734e39a45db213c up
that sample test classes are stored in a specific container.
